NETUSER: specify the type of networking
Allow the user to specify the type of networking:
1) simple, outgoing only, no ping -> "-net user"
2) bidirectional, full networking
   - requires qnet_up/qnet_down
   - dns server
   - tap device
   - iptables configuration
   - routing adjustments

By default option 2) will get selected (i.e. the more complicated version,
making this type of networking easy was the primary motivation for this
project!). If the user just wants plain output-only networking, simply add the
following to the image's CONFIG file:

	NETUSER=1

Specifying zero or leaving this option out altogether selects the default.

Signed-off-by: Trevor Woerner <twoerner@gmail.com>