USEGUI: fix when set to zero
If the user doesn't specify the "USEGUI" option, then the cmdline will be used
by default. This commit adds the case that the user has provided the "USEGUI"
option but has set it to zero.

Previously, providing this option in any way (even setting it to zero) would
cause the non-default behaviour.

Signed-off-by: Trevor Woerner <twoerner@gmail.com>