qnet_up/down: dhcp/dns fixups
While working with a new machine (that has a much more restrictive firewall,
and one that I can't simply turn the firewall off) I noticed that the dhcp/dns
stuff wasn't working because more holes needed to be punched through the
firewall.

I also noticed that:
- the host had no problem doing DNS resolution
- the client had no problem talking to any machine
- by default dnsmasq sets the nameserver for the client to the host (itself)
- for some reason DNS resolution doesn't succeed

So if the client were told of the upstream DNS server then it would be able to
use it to perform successful DNS queries. But when it tries to use the host
for DNS, it fails. This is the purpose of enabling --dhcp-option #6. This
tells dnsmasq to tell the client to use a specific IP address for
/etc/resolv.conf (i.e. the one that it itself is using) instead of using
itself (i.e. the host).

Signed-off-by: Trevor Woerner <twoerner@gmail.com>