qnet_up/down: better dnsmasq pid handling
If the dnsmasq app fails to start, make sure the root user doesn't try to kill
either some non-existent process (best case) or some random process (worse
case) when the VM is terminated (which could be days/weeks/etc later!). The
handling involves looking for the dnsmasq process immediately after starting
it, and only writing the *pid file if it can be found.

Signed-off-by: Trevor Woerner <twoerner@gmail.com>